resolved problems caused by targeting wrong .net framework

+++ SubnauticaWatcherMod/Patching/Patcher.cs
@@ -42,13 +42,13 @@
 
                 // Yes, Z and Y are supposed to be reversed. Converting Subnautica coords to something more traditional.
                 PlayerInfo.Instance.X = (int) Math.Round(
-                    __instance.transform.position.x,
+                    (double)__instance.transform.position.x,
                     MidpointRounding.AwayFromZero);
                 PlayerInfo.Instance.Z = (int) Math.Round(
-                    __instance.transform.position.y,
+                    (double)__instance.transform.position.y,
                     MidpointRounding.AwayFromZero);
                 PlayerInfo.Instance.Y = (int) Math.Round(
-                    __instance.transform.position.z,
+                    (double)__instance.transform.position.z,
                     MidpointRounding.AwayFromZero);
             }
         }
